app/requirements.txt
fastapi
uvicorn
requests
openai
pypdf
python-multipart
bs4
faiss-cpu
weasyprint
jinja2